% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_gib_read.R
\name{full_gib_read}
\alias{full_gib_read}
\title{Wrapped read in functionality}
\usage{
full_gib_read(path = NULL)
}
\arguments{
\item{path}{path to smart survey data}
}
\value{
clean GiB dataset
}
\description{
Wrapped read in functionality
}
\examples{
\dontrun{
full_gib_read("23-03-27.csv")
}
}
