% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gib_clean_data.R
\name{gib_clean_data}
\alias{gib_clean_data}
\title{Join all gib datasets}
\usage{
gib_clean_data(
  gib_raw_user_details = NULL,
  gib_raw_character_timber_details = NULL,
  gib_raw_gps_timber_details = NULL,
  gib_raw_numeric_timber_details = NULL
)
}
\arguments{
\item{gib_raw_user_details}{parsed user details}

\item{gib_raw_character_timber_details}{parsed char timber details}

\item{gib_raw_gps_timber_details}{parsed gps timber details}

\item{gib_raw_numeric_details}{parsed num timber details}
}
\value{
clean GiB dataset
}
\description{
Join all gib datasets
}
\examples{
\dontrun{
gib_clean_data(gib_raw)
}
}
