% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hwpsc-package.R
\docType{package}
\name{hwpsc-package}
\alias{hwpsc}
\alias{hwpsc-package}
\title{hwpsc: Analyse Hardwood Price Size Curves in R}
\description{
An R package with code and documentation for the analysis of hardwood pricing as part of The Economics of Woodland Creation project.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fr-dan/hwpsc}
  \item Report bugs at \url{https://github.com/fr-dan/hwpsc/issues}
}

}
\author{
\strong{Maintainer}: 

Other contributors:
\itemize{
  \item Daniel Braby [contributor]
  \item Jack Forster [contributor]
}

}
\keyword{internal}
