% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_char_timber_details.R,
%   R/parse_gps_timber_details.R, R/parse_num_timber_details.R,
%   R/parse_timber_details.R, R/parse_user_details.R
\name{parse_char_timber_details}
\alias{parse_char_timber_details}
\alias{parse_gps_timber_details}
\alias{parse_num_timber_details}
\alias{parse_timber_details}
\alias{parse_user_details}
\title{Parse character details from timber details GiB data}
\usage{
parse_char_timber_details(df = NULL)

parse_gps_timber_details(df = NULL)

parse_num_timber_details(df = NULL)

parse_timber_details(df = NULL)

parse_user_details(df = NULL)
}
\arguments{
\item{df}{data.frame with raw GiB survey data}
}
\value{
GiB character timber details in data.frame

GiB GPS timber details in data.frame

GiB numeric timber details in data.frame

GiB timber details in data.frame

GiB user details in data.frame
}
\description{
Parse character details from timber details GiB data

Parse GPS details from timber details GiB data

Parse numeric details from timber details GiB data

Parse timber details from raw GiB survey data

Parse user details from raw GiB survey data
}
\examples{
\dontrun{
parse_char_timber_details(gib_raw)
}
\dontrun{
parse_gps_timber_details(gib_raw)
}
\dontrun{
parse_num_timber_details(gib_raw)
}
\dontrun{
parse_timber_details(gib_raw)
}
\dontrun{
parse_user_details(gib_raw)
}
}
