% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gib_raw.R
\name{read_gib_raw}
\alias{read_gib_raw}
\title{Read in raw GiB Smart Survey data}
\usage{
read_gib_raw(path = NULL)
}
\arguments{
\item{path}{file path}
}
\value{
raw GiB survey data
}
\description{
Read in raw GiB Smart Survey data
}
\examples{
\dontrun{
read_gib_raw(here("data_raw", "23-03-27.csv"))
}
}
